+++ Vic2ToHoI4/Source/HOI4World/HoI4Country.h
@@ -87,7 +87,7 @@ class HoI4Country
 		void addState(HoI4State* _state);
 		void		lowerNeutrality(double amount);
 		void calculateIndustry();
-		void addVPsToCapital(double VPs);
+		void addVPsToCapital(int VPs);
 
 		void		setSphereLeader(string SphereLeader) { sphereLeader == SphereLeader; }
 		void		setFaction(string newFaction)	{ faction = newFaction; }
